refac(91026): Extrai task concluir_prestacao_de_contas_async
