+++ sme_ptrf_apps/core/tasks/__init__.py
@@ -1,1 +1,3 @@
 from .tasks_legado import *
+
+from .concluir_prestacao_de_contas import concluir_prestacao_de_contas_async
